refactor: Extract Difficulty and Category into their own component files for reuse
