+++ src/styles/compiled.css
@@ -1783,11 +1783,11 @@ select[multiple]:focus option:checked {
 .mt-3 {
   margin-top: 0.75rem;
 }
-.mr-3 {
-  margin-right: 0.75rem;
-}
 .ml-3 {
   margin-left: 0.75rem;
+}
+.mr-3 {
+  margin-right: 0.75rem;
 }
 .mb-3 {
   margin-bottom: 0.75rem;
@@ -1953,11 +1953,11 @@ select[multiple]:focus option:checked {
 .max-w-\[60rem\] {
   max-width: 60rem;
 }
-.max-w-\[420px\] {
-  max-width: 420px;
-}
 .max-w-\[90\%\] {
   max-width: 90%;
+}
+.max-w-\[420px\] {
+  max-width: 420px;
 }
 .max-w-\[350px\] {
   max-width: 350px;
